quantum-adapter-kis/pyproject.toml
[project]
name = "kis-github"
version = "1.0.0"
description = "korea investment open api sample codes"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.3.1",
    "pycryptodome>=3.23.0",
    "pyqt6>=6.9.1",
    "pyside6>=6.9.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "websockets>=15.0.1",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "backtrader>=1.9.78.123",
    "pandas-ta>=0.3.14b0",
    "quantstats>=0.0.76",
    "matplotlib>=3.10.6",
    # 추가 데이터 소스 라이브러리 (KIS API 중심화로 제거)
    # "pykrx>=1.0.51",           # 제거: KIS API로 대체
    # "finance-datareader>=0.9.0", # 제거: KIS API로 대체
    # "yfinance>=0.2.0",          # 제거: KIS API로 대체
    # 고성능 백테스팅 라이브러리 (조건부)
    # "vectorbt>=0.26.1",  # 별도 설치
    # 기술적 분석 라이브러리 (조건부)
    # "TA-Lib>=0.4.28",  # 별도 설치
    # 추가 유틸리티
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "psycopg2-binary>=2.9.10",
    "httpx>=0.28.1",
    "pyjwt>=2.10.1",
    "openai>=1.107.0",
]
